Update README.md
HOW TO USE THE APPLICATION
----------------------------
First of download all the files.  (ALL THE FILES ARE NECESSARY FOR IT TO WORK SUCCESSFULLY.)
There are 2 buttons with image of a WAV and Mp3.  The Mp3 will convert your video file to .MP3 audio file and same wav will convert it to wav file.
Click the screen recorder button to start recording the screen.  Once finished press 'esc' button of your keyboard after opening the screen recorder tab.